feat(#S11P12D203-227): studyId 데이터 타입 Long으로 변경

+++ android/project/app/src/main/java/com/chill/mallang/data/model/response/IncorrectWordResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class IncorrectWordResponse(
     @SerialName("studyId")
-    val studyId: Int? = null,
+    val studyId: Long? = null,
     @SerialName("word")
     val script: String? = ""
 )
